HUB: refactor GameInfo into its own file (no functional change)
